tech(api): move getAnswer usecase
